Initial implementation of reject page
