+++ src/views/supplier/manage-shares.njk
@@ -15,7 +15,7 @@
   <div class="govuk-grid-column-one-half">
     <h2 class="govuk-heading-l">Created requests</h2>
     <p class="govuk-body">View pending, submitted and completed requests to access data owned by others.</p>
-    <p class="govuk-body"><a href="manage-shares/created-requests" class="govuk-link">View created requests</a></p>
+    <p class="govuk-body"><a href="acquirer/created-requests" class="govuk-link">View created requests</a></p>
   </div>
 </div>
 {% endblock %}
