+++ src/views/supplier/received-requests.njk
@@ -13,7 +13,7 @@
     <aside class="app-related-items">
         <nav role="navigation" aria-labelledby="subsection-title">
           <ul class="govuk-list">
-            <li><a href="created-requests" class="govuk-link">View created requests</a></li>
+            <li><a href="{{assetPath}}/acquirer/created-requests" class="govuk-link">View created requests</a></li>
           </ul>
         </nav>
     </aside>
